Change time and distance calculation from long to short in routing connections
The on disk file format only used shorts for these variables anyway, so no need to use
larger variables during calculation. Also include checks to see if we overflow.

+++ Osm2GpsMid/src/de/ueller/osmToGpsMid/AStar.java
@@ -164,7 +164,7 @@ public class AStar {
 	}
 	// {{{ public final Vector solve (State initialState) 
 	public final Vector<Connection> solve (RouteNode start,RouteNode target) {
-		Connection initialState=new Connection(start,0l,0l,(byte)0,(byte)0,null);
+		Connection initialState=new Connection(start,(short)0,(short)0,(byte)0,(byte)0,null);
 		Node solution;
 		Node firstNode;
 		long estimation;
