+++ Osm2GpsMid/src/de/ueller/osmToGpsMid/AStar2.java
@@ -249,7 +249,7 @@ public class AStar2 {
 	public final Vector<Connection> solve (RouteNode start,RouteNode target) {
 		long totalDist = MyMath.dist(start.node, target.node,1.2);
 		long estimateSpeed=0;
-		Connection initialState=new Connection(start,0l,0l,(byte)0,(byte)0,null);
+		Connection initialState=new Connection(start,(short)0,(short)0,(byte)0,(byte)0,null);
 		Node solution;
 		Node firstNode;
 		long estimation;
